Turbopack: source locations as u32 (#75705)
No real signficant perf difference, but seems still worthwhile. Especially since the `sourcemap` library and `swc` both return u32 anyway. And you likely don't have files with more than 4.294.967.296 bytes, lines, or column width.

```
snapshot in with_modules c2c661b1e9 (previous PR in stack)
13,3 GB
TURBO_ENGINE_READ_ONLY=1 NEXT_TURBOPACK_TRACING= TURBOPACK=1 TURBOPACK_BUILD=  491.50s user 79.78s system 856% cpu 1:06.69 total
TURBO_ENGINE_READ_ONLY=1 NEXT_TURBOPACK_TRACING= TURBOPACK=1 TURBOPACK_BUILD=  495.62s user 78.61s system 852% cpu 1:07.36 total

locations as u32 b422f84b24
13,25 GB
TURBO_ENGINE_READ_ONLY=1 NEXT_TURBOPACK_TRACING= TURBOPACK=1 TURBOPACK_BUILD=  494.85s user 77.34s system 855% cpu 1:06.87 total
```

Closes PACK-3940

+++ crates/napi/src/next_api/project.rs
@@ -1190,8 +1190,8 @@ pub async fn project_trace_source(
 
             let token = map
                 .lookup_token(
-                    (line as usize).saturating_sub(1),
-                    (frame.column.unwrap_or(1) as usize).saturating_sub(1),
+                    line.saturating_sub(1),
+                    frame.column.unwrap_or(1).saturating_sub(1),
                 )
                 .await?;
 
@@ -1199,8 +1199,8 @@ pub async fn project_trace_source(
                 Token::Original(token) => (
                     urlencoding::decode(&token.original_file)?.into_owned(),
                     // JS stack frames are 1-indexed, source map tokens are 0-indexed
-                    Some(token.original_line as u32 + 1),
-                    Some(token.original_column as u32 + 1),
+                    Some(token.original_line + 1),
+                    Some(token.original_column + 1),
                     token.name.clone(),
                 ),
                 Token::Synthetic(token) => {
